==> 2_write_json_file_to_csv.ipynb <==
import os, sys, csv, json, re
import numpy as np
import pandas as pd
import datetime as dt

## Read run time JSON

scene_folder = 'scene-json'
case_folder = '2-cityffd-97' ## todo - JSON folder
header = 'sf_100_dt_hvac_cfd_r2_09_' ## todo - JSON header
CSV_FILE_NAME = 'sf_97_2m_hvac_cfd_r2.csv' ## todo - CSV Schedule file header
hourly_data = {}
for day in range(2, 0, -1):
    for hour in range(1, 25):
#         print(os.path.join(scene_folder, case_folder, header + str(day).zfill(2) + '_' + str(hour).zfill(2) + '_00.json'))
        with open(os.path.join(scene_folder, case_folder, header + str(day).zfill(2) + '_' + str(hour).zfill(2) + '_00.json'), 'r') as f:
            wrf_data = json.load(f)
            for wrf_grid in wrf_data:
                wrf_id = wrf_grid['id']
                dry_bulb = wrf_grid['properties']['dry_bulb_temperature']
                wind_speed = wrf_grid['properties']['wind_speed']
                wind_direction = wrf_grid['properties']['wind_direction']
                if wrf_id not in hourly_data:
                    hourly_data[wrf_id] = [(dry_bulb, wind_speed, wind_direction)]
                else:
                    hourly_data[wrf_id].append((dry_bulb, wind_speed, wind_direction))

hourly_data[1][:12]

## Write 48 hour JSON to 48 hour CSV

CSV_FILE_NAME = 'sf_97_2m_hvac_cfd_r2_day.csv' ## todo
with open(os.path.join(scene_folder, 'run_time_inputs', CSV_FILE_NAME), 'w') as out:
    header_line = []
    for wrf_id in hourly_data:      
        for header in ['drybulb', 'wind_speed', 'wind_dir']:
            header_name = header + ':' + str(wrf_id)
            header_line.append(header_name)
    out.write(','.join(header_line) + '\n')
    for hour in range(24, 48):
        curr_line = []
        for wrf_id in hourly_data:
            curr_line.append(str(round(hourly_data[wrf_id][hour][0], 3)))
            curr_line.append(str(round(hourly_data[wrf_id][hour][1], 3)))
            curr_line.append(str(round(hourly_data[wrf_id][hour][2], 3)))
        out.write(','.join(curr_line) + '\n')
    for hour in range(24):
        curr_line = []
        for wrf_id in hourly_data:
            curr_line.append(str(round(hourly_data[wrf_id][hour][0], 3)))
            curr_line.append(str(round(hourly_data[wrf_id][hour][1], 3)))
            curr_line.append(str(round(hourly_data[wrf_id][hour][2], 3)))
        out.write(','.join(curr_line) + '\n')
        
            

## Write 48 hour JSON to 8760 hour CSV

CSV_FILE_NAME = 'sf_97_2m_hvac_cfd_r2.csv'
with open(os.path.join(scene_folder, 'run_time_inputs', CSV_FILE_NAME), 'w') as out:
    header_line = []
    for wrf_id in hourly_data:      
        for header in ['drybulb', 'wind_speed', 'wind_dir']:
            header_name = header + ':' + str(wrf_id)
            header_line.append(header_name)
    out.write(','.join(header_line) + '\n')
    for day in range(365 // 2):
        for hour in range(48):
            curr_line = []
            for wrf_id in hourly_data:
                curr_line.append(str(round(hourly_data[wrf_id][hour][0], 3)))
                curr_line.append(str(round(hourly_data[wrf_id][hour][1], 3)))
                curr_line.append(str(round(hourly_data[wrf_id][hour][2], 3)))
            out.write(','.join(curr_line) + '\n')
    for hour in range(24):
        curr_line = []
        for wrf_id in hourly_data:
            curr_line.append(str(round(hourly_data[wrf_id][hour][0], 3)))
            curr_line.append(str(round(hourly_data[wrf_id][hour][1], 3)))
            curr_line.append(str(round(hourly_data[wrf_id][hour][2], 3)))
        out.write(','.join(curr_line) + '\n')

SCHEDULE_FILE_NAME = 'sf_97_2m_hvac_cfd_r2.csv'
df_csv_result = pd.read_csv(os.path.join(scene_folder, 'run_time_inputs', SCHEDULE_FILE_NAME), sep=',', encoding='UTF-8')
df_csv_result.head()

## Separate to smaller files for faster reading

for i in range(34):
    new_result = pd.DataFrame()
    for j in range(1, 101):
        grid_id = i * 100 + j
        for header in ['drybulb', 'wind_speed', 'wind_dir']:
            if not header + ':' + str(grid_id) in df_csv_result:
                continue
            new_result[header + ':' + str(grid_id)] = df_csv_result[header + ':' + str(grid_id)]
    new_result.to_csv(os.path.join(scene_folder, 'run_time_inputs', "sf_97_2m_hvac_cfd_r2_" + str(i) + '.csv')) # todo
